cleaned up UI and added a tournament selector to the weekly picks page
