+++ client-web/src/pages/weeklyPicks/index.jsx
@@ -9,7 +9,6 @@ import FullTournamentTable from 'src/assets/components/fullTournamentTable';
 const WeeklyPicks = () => {
 
   const [activeSeason, setActiveSeason] = useState(null);
-  const [isActiveTournament, setIsActiveTournament] = useState(true);
   const [activeTournament, setActiveTournament] = useState(null);
   
   const { isLoggedIn } = useSelector(state => state.auth);
@@ -27,33 +26,11 @@ const WeeklyPicks = () => {
     }
   }, [isLoggedIn]);
 
-  useEffect(() => {
-    if (activeSeason !== null) {
-      SeasonsApi.activeTournament(activeSeason.id).then(
-        (response) => {
-          if (response.status === 200) {
-            setIsActiveTournament(true);
-            setActiveTournament(response.data);
-          }
-          if (response.status === 204) {
-            setIsActiveTournament(false);
-          }
-        },
-        (error) => error
-      );
-    }
-  }, [activeSeason]);
-
   return (
     <Box sx={{ flexGrow: 1, textAlign: 'center'}}>
       <Grid container className='mt-3 px-2'>
         <Grid item xs={12}>
-      {
-        isActiveTournament ? 
-        (activeSeason && activeTournament && <FullTournamentTable season={activeSeason} tournament={activeTournament}/>)
-      :
-        <h1>There is no currently active tournament. Please check in again while there is an ongoing event!</h1>
-      }
+          <FullTournamentTable seasonId={activeSeason?.id}/>
         </Grid>
       </Grid>
     </Box>
